[tests] generator-Tests pass on Windows
- `CSharpCodeProvider` needs a `using` or the generated assembly file is
locked for future tests. This causes many test failures.
- `CSharpCodeProvider` won't support C# 6 on Windows, using NuGet
package for Rosyln instead
- The Roslyn package seems to require an environment variable to locate
`csc.exe`. See: https://github.com/aspnet/RoslynCodeDomProvider/pull/12
- When running on Mono, the use of the `partial` keyword is generated as
`partial_`. This does not happen on Windows, so included a set of
"expected" files for Windows. This runs a different file comparison if
the tests are running on Windows.