[generator] integration tests use multiple assemblies
`BaseGeneratorTest` should put `Java.Lang.*` types in a separate
assembly, in order to properly test what happens to a user's Java
binding project. We need to write these types in a separate assembly
to disk, so that the main in-memory assembly under test can reference
it.

This will also help fix many of the compiler warnings down the road
where the `new` keyword is being emitted when it shouldn't be.

Changes to support this:
- `BaseGeneratorTest` can handle additional support files for the
  "support" assembly OR the "main" assembly
- `Streams` & `Adapters` tests need to use additional "source" files
- `JavaConvert` is now `public`, so support files compile for
  `Adapters` class
- `Java_Lang_Object` test needs to load support assembly to check
  `Java.Lang.Object`
- Wrote an `AcrossAssemblies` test to just ensure basic things are
  working when inlcuding additional types in the "support" assembly.

Additionally to make sure we don't lose anything, I added a test that
compiles everything to a single assembly the way `Mono.Android.csproj`
does in `xamarin-android`. For now, I included a single class for
`android.app.Activity` pulled from API 27 to verify that generator's
output works and compiles.